refactored app into homepage
