+++ src/components/PlantCard/PlantCard.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import PlantDetails from '../PlantDetails/PlantDetails.jsx';
-import fetchPlantDetails from '../PlantDetails/fetchPlantDetails.jsx';
+import fetchPlantDetails from '../PlantDetails/fetchPlantDetailsApi.jsx';
 import { auth, db } from "../../firebase/firebase.js";
 import { collection, addDoc, getDocs, serverTimestamp } from "firebase/firestore";
 import { useAuth } from "../Authentication/AuthContext";
@@ -11,7 +11,7 @@ import styles from "./PlantCard.module.css";
 const PlantCard = ({plant, onDelete, isDashboard}) => {
 const [showPlantDetail, setShowPlantDetail] = useState(false);
 const navigate = useNavigate();
-const { userLoggedIn, currentUser } = useAuth();
+const { currentUser } = useAuth();
 
 const handleDetailsClick = () => {
 	setShowPlantDetail(prev=>!prev);
