+++ src/components/PlantDetails/fetchPlantDetailsApi.jsx
